agent-driven-w-tests/antigravity-llm-logic-eval/requirements.txt
python-constraint==1.4.0
openai==2.22.0
